requirements.txt
alembic==1.13.1
black==24.4.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
flake8==7.0.0
Flask==3.0.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pipp==0.0.1
platformdirs==4.2.1
pycodestyle==2.11.1
pyflakes==3.2.0
PyMySQL==1.1.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.18.1
